GEM-18843: Update versions
- com.github.ben-manes.versions: 0.50.0 -> 0.54.0
- gemfireVersion: 10.2.0 -> 10.2.4
- gradle: 9.4.1 -> 9.5.1
- jakarta.servlet.jsp.jstl-api: 3.0.0 -> 3.0.2
- kotlin-gradle-plugin: 2.1.20 -> 2.2.21
- logback-classic: 1.5.12 -> 1.5.33
- nl.littlerobots.version-catalog-update: 0.8.5 -> 1.1.0
- slf4j: 2.0.16 -> 2.0.18
- spring-framework: 6.2.12 -> 6.2.18
- spring-security: 6.5.6 -> 6.5.10
- spring-session: 3.5.3 -> 3.5.6

ai-assisted=yes
