+++ build-tools/publishing/src/main/kotlin/gemfire-repo-artifact-publishing.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2025 Broadcom. All rights reserved.
+ * Copyright 2023-2026 Broadcom. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
 
@@ -16,7 +16,7 @@ plugins {
 project.ext.set("constrainVersionInBom", true)
 
 val license = """
-  Copyright (c) VMware, Inc. 2023-2025. All rights reserved.
+  Copyright (c) VMware, Inc. 2023-2026. All rights reserved.
   SPDX-License-Identifier: Apache-2.0
   """
