+++ gradle.properties
@@ -1,10 +1,11 @@
 #
-# Copyright 2022-2024 Broadcom. All rights reserved.
+# Copyright 2022-2026 Broadcom. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 springDataGemFireVersion=1.0.0
 version=1.0.0-build.9999
-gemfireVersion=10.2.0
+gemfireVersion=10.2.4
+baseGemFireVersion=10.2
 
 mavenPushRepository=
 org.gradle.workers.max=3
