+++ gradle/publish.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-kotlinGradlePluginVersion = "2.1.20"
+kotlinGradlePluginVersion = "2.2.21"
 
 
 [libraries]
