+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,10 +1,10 @@
 #
-# Copyright 2024 Broadcom. All rights reserved.
+# Copyright 2024-2026 Broadcom. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-9.4.1-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-9.5.1-bin.zip
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
